requirements.txt
pytest==8.3.2
pytest-html==4.1.1
pytest-xdist==3.6.1
selenium==4.23.1
webdriver-manager==4.0.2
allure-pytest==2.15.0
requests==2.32.3
